Return immutable slices
The use case for mutable slices just wasn't there. One would likely not
want to modify the ciphertext. The plaintext, maybe, but if that's the
case, it's easy for the user to do, since they own the underlying data.

And by principle, if we can get away with an immutable slice, we should
do so, as that allows the presence of several at the same time.