doc: Be explicit about parameters, and link COSE algorithms
While the parameters of AES-CCM are encoded in the types of the NewAead
and Aead implementations, those are not visible by default and sometimes
not given in the common unit of bits. This adds explicit text
documentation that makes those visible.

In addition, the equivalent COSE algorithms are stated; this is useful
because their parameterization is in terms of log(maximum message size),
while the Aead parameters are given in terms of nonce size. Mentioning
the two relevant versions explicitly makes them easier to discover in
generic full-text search.